Add the option to shortcut eproject file searching if project already open.
When working remotely over something like TRAMP performance can be
severely impacted as eproject checks against all possible eproject
types. This optimisation will assume that if a file is opened in a
subdirectory of an existing loaded project then we can use that
definition.

This breaks sub-projects although this could be mitigated if we sort
the existing roots by length and the user opens sub-projects before
the main project.